+ words to __m128i, program bacome faster, now it need 9,7 s, while in standart second version was 10,4 s

+++ main.cpp
@@ -31,7 +31,7 @@ int main (int argc, char* argv[])
 
         if (argc > 3)
         {
-            ProcessMeasurings (&hashtable, NTIMES, atoi (argv[3]));
+            ProcessMeasurings (&hashtable, NTIMES, (size_t) atoi (argv[3]));
         }
         else
         {
